+++ CHANGES.md
@@ -6,6 +6,7 @@
 -   Added a function to update the request count of an organisation to match
     its completed tasks.
 -   Check status view now also correctly updates the status
+-   Added Chloride to fieldwork options
 
 
 ## 1.57 (2025-11-12)